Making ReachHover "dumb-mode" safe
When IntelliJ is in "dumb-mode" (unfortunate phrasing), most of its static
analysis facilities are not available. In fact, invoking most features will
lead to the IDE throwing an `IndexNotReadyException`.

Previously, `reach-hover` was not respecting this. Now, it displays a warning
to the end-user stating that it can be invoked after background tasks, such
as indexing, are complete.