Improving notification performance
Previously, we were issuing a notification every time the user invoked
`reach-hover` code when IntelliJ was running background tasks.

This is incredibly inefficient due to the high sample rate at which listeners
poll for events.

We now only issue `reach-hover` "not ready" warnings when the previous
notification has been expired.